Added version 2.5.1525

+++ chromium/build/android/gyp/util/resource_utils.py
@@ -25,22 +25,127 @@ EMPTY_ANDROID_MANIFEST_PATH = os.path.join(
     _SOURCE_ROOT, 'build', 'android', 'AndroidManifest.xml')
 
 
-# A variation of this lists also exists in:
+# A variation of these maps also exists in:
 # //base/android/java/src/org/chromium/base/LocaleUtils.java
 # //ui/android/java/src/org/chromium/base/LocalizationUtils.java
-CHROME_TO_ANDROID_LOCALE_MAP = {
-    'en-GB': 'en-rGB',
-    'en-US': 'en-rUS',
+_CHROME_TO_ANDROID_LOCALE_MAP = {
     'es-419': 'es-rUS',
     'fil': 'tl',
     'he': 'iw',
     'id': 'in',
-    'pt-PT': 'pt-rPT',
-    'pt-BR': 'pt-rBR',
     'yi': 'ji',
-    'zh-CN': 'zh-rCN',
-    'zh-TW': 'zh-rTW',
 }
+_ANDROID_TO_CHROMIUM_LANGUAGE_MAP = {
+    'tl': 'fil',
+    'iw': 'he',
+    'in': 'id',
+    'ji': 'yi',
+    'no': 'nb',  # 'no' is not a real language. http://crbug.com/920960
+}
+
+
+def ToAndroidLocaleName(chromium_locale):
+  """Convert an Chromium locale name into a corresponding Android one."""
+  # First handle the special cases, these are needed to deal with Android
+  # releases *before* 5.0/Lollipop.
+  android_locale = _CHROME_TO_ANDROID_LOCALE_MAP.get(chromium_locale)
+  if android_locale:
+    return android_locale
+
+  # Format of Chromium locale name is '<lang>' or '<lang>-<region>'
+  # where <lang> is a 2 or 3 letter language code (ISO 639-1 or 639-2)
+  # and region is a capitalized locale region name.
+  lang, _, region = chromium_locale.partition('-')
+  if not region:
+    return lang
+
+  # Translate newer language tags into obsolete ones. Only necessary if
+  #  region is not None (e.g. 'he-IL' -> 'iw-rIL')
+  lang = _CHROME_TO_ANDROID_LOCALE_MAP.get(lang, lang)
+
+  # Using '<lang>-r<region>' is now acceptable as a locale name for all
+  # versions of Android.
+  return '%s-r%s' % (lang, region)
+
+
+# ISO 639 language code + optional ("-r" + capitalized region code).
+# Note that before Android 5.0/Lollipop, only 2-letter ISO 639-1 codes
+# are supported.
+_RE_ANDROID_LOCALE_QUALIFIER_1 = re.compile(r'^([a-z]{2,3})(\-r([A-Z]+))?$')
+
+# Starting with Android 7.0/Nougat, BCP 47 codes are supported but must
+# be prefixed with 'b+', and may include optional tags. e.g. 'b+en+US',
+# 'b+ja+Latn', 'b+ja+JP+Latn'
+_RE_ANDROID_LOCALE_QUALIFIER_2 = re.compile(r'^b\+([a-z]{2,3})(\+.+)?$')
+
+# Matches an all-uppercase region name.
+_RE_ALL_UPPERCASE = re.compile(r'^[A-Z]+$')
+
+
+def ToChromiumLocaleName(android_locale):
+  """Convert an Android locale name into a Chromium one."""
+  lang = None
+  region = None
+  m = _RE_ANDROID_LOCALE_QUALIFIER_1.match(android_locale)
+  if m:
+    lang = m.group(1)
+    if m.group(2):
+      region = m.group(3)
+  else:
+    m = _RE_ANDROID_LOCALE_QUALIFIER_2.match(android_locale)
+    if m:
+      lang = m.group(1)
+      if m.group(2):
+        tags = m.group(2).split('+')
+        # First all-uppercase tag is a region. This deals with cases where
+        # a special tag is placed before it (e.g. 'cmn+Hant-TW')
+        for tag in tags:
+          if _RE_ALL_UPPERCASE.match(tag):
+            region = tag
+            break
+
+  if not lang:
+    return None
+
+  # Special case for es-rUS -> es-419
+  if lang == 'es' and region == 'US':
+    return 'es-419'
+
+  lang = _ANDROID_TO_CHROMIUM_LANGUAGE_MAP.get(lang, lang)
+  if not region:
+    return lang
+
+  return '%s-%s' % (lang, region)
+
+
+def IsAndroidLocaleQualifier(string):
+  """Returns true if |string| is a valid Android resource locale qualifier."""
+  return (_RE_ANDROID_LOCALE_QUALIFIER_1.match(string)
+          or _RE_ANDROID_LOCALE_QUALIFIER_2.match(string))
+
+
+def FindLocaleInStringResourceFilePath(file_path):
+  """Return Android locale name of a string resource file path.
+
+  Args:
+    file_path: A file path.
+  Returns:
+    If |file_path| is of the format '.../values-<locale>/<name>.xml', return
+    the value of <locale> (and Android locale qualifier). Otherwise return None.
+  """
+  if not file_path.endswith('.xml'):
+    return None
+  prefix = 'values-'
+  dir_name = os.path.basename(os.path.dirname(file_path))
+  if not dir_name.startswith(prefix):
+    return None
+  qualifier = dir_name[len(prefix):]
+  return qualifier if IsAndroidLocaleQualifier(qualifier) else None
+
+
+def ToAndroidLocaleList(locale_list):
+  """Convert a list of Chromium locales into the corresponding Android list."""
+  return sorted(ToAndroidLocaleName(locale) for locale in locale_list)
 
 # Represents a line from a R.txt file.
 _TextSymbolEntry = collections.namedtuple('RTextEntry',
@@ -68,7 +173,7 @@ def _ParseTextSymbolsFile(path, fix_package_ids=False):
 
   Args:
     path: Input file path.
-    fix_package_ids: if True, all packaged IDs read from the file
+    fix_package_ids: if True, 0x00 and 0x02 package IDs read from the file
       will be fixed to 0x7f.
   Returns:
     A list of _TextSymbolEntry instances.
@@ -92,20 +197,49 @@ def _FixPackageIds(resource_value):
   # Resource IDs for resources belonging to regular APKs have their first byte
   # as 0x7f (package id). However with webview, since it is not a regular apk
   # but used as a shared library, aapt is passed the --shared-resources flag
-  # which changes some of the package ids to 0x02 and 0x00.  This function just
-  # normalises all package ids to 0x7f, which the generated code in R.java
-  # changes to the correct package id at runtime.
+  # which changes some of the package ids to 0x00 and 0x02.  This function
+  # normalises these (0x00 and 0x02) package ids to 0x7f, which the generated
+  # code in R.java changes to the correct package id at runtime.
   # resource_value is a string with either, a single value '0x12345678', or an
   # array of values like '{ 0xfedcba98, 0x01234567, 0x56789abc }'
-  return re.sub(r'0x(?!01)\d\d', r'0x7f', resource_value)
+  return re.sub(r'0x(?:00|02)', r'0x7f', resource_value)
 
 
 def _GetRTxtResourceNames(r_txt_path):
   """Parse an R.txt file and extract the set of resource names from it."""
-  result = set()
-  for entry in _ParseTextSymbolsFile(r_txt_path):
-    result.add(entry.name)
-  return result
+  return {entry.name for entry in _ParseTextSymbolsFile(r_txt_path)}
+
+
+def GetRTxtStringResourceNames(r_txt_path):
+  """Parse an R.txt file and the list of its string resource names."""
+  return sorted({
+      entry.name
+      for entry in _ParseTextSymbolsFile(r_txt_path)
+      if entry.resource_type == 'string'
+  })
+
+
+def GenerateStringResourcesWhitelist(module_r_txt_path, whitelist_r_txt_path):
+  """Generate a whitelist of string resource IDs.
+
+  Args:
+    module_r_txt_path: Input base module R.txt path.
+    whitelist_r_txt_path: Input whitelist R.txt path.
+  Returns:
+    A dictionary mapping numerical resource IDs to the corresponding
+    string resource names. The ID values are taken from string resources in
+    |module_r_txt_path| that are also listed by name in |whitelist_r_txt_path|.
+  """
+  whitelisted_names = {
+      entry.name
+      for entry in _ParseTextSymbolsFile(whitelist_r_txt_path)
+      if entry.resource_type == 'string'
+  }
+  return {
+      int(entry.value, 0): entry.name
+      for entry in _ParseTextSymbolsFile(module_r_txt_path)
+      if entry.resource_type == 'string' and entry.name in whitelisted_names
+  }
 
 
 class RJavaBuildOptions:
@@ -189,9 +323,8 @@ class RJavaBuildOptions:
       return entry.name not in self.resources_whitelist
 
 
-def CreateRJavaFiles(srcjar_dir, package, main_r_txt_file,
-                     extra_res_packages, extra_r_txt_files,
-                     rjava_build_options):
+def CreateRJavaFiles(srcjar_dir, package, main_r_txt_file, extra_res_packages,
+                     extra_r_txt_files, rjava_build_options):
   """Create all R.java files for a set of packages and R.txt files.
 
   Args:
@@ -329,7 +462,11 @@ public final class R {
         public static final {{ e.java_type }} {{ e.name }} = {{ e.value }};
         {% endfor %}
         {% for e in non_final_resources[resource_type] %}
+            {% if e.value != '0' %}
         public static {{ e.java_type }} {{ e.name }} = {{ e.value }};
+            {% else %}
+        public static {{ e.java_type }} {{ e.name }};
+            {% endif %}
         {% endfor %}
     }
     {% endfor %}
@@ -383,8 +520,8 @@ def ExtractDeps(dep_zips, deps_dir):
 
   Args:
      dep_zips: A list of zip file paths, each one will be extracted to
-       a subdirectory of |deps_dir|, named after the zip file (e.g.
-       '/some/path/foo.zip' -> '{deps_dir}/foo/').
+       a subdirectory of |deps_dir|, named after the zip file's path (e.g.
+       '/some/path/foo.zip' -> '{deps_dir}/some_path_foo/').
     deps_dir: Top-level extraction directory.
   Returns:
     The list of all sub-directory paths, relative to |deps_dir|.
@@ -394,20 +531,32 @@ def ExtractDeps(dep_zips, deps_dir):
   """
   dep_subdirs = []
   for z in dep_zips:
-    subdir = os.path.join(deps_dir, os.path.basename(z))
+    subdirname = z.replace(os.path.sep, '_')
+    subdir = os.path.join(deps_dir, subdirname)
     if os.path.exists(subdir):
-      raise Exception('Resource zip name conflict: ' + os.path.basename(z))
+      raise Exception('Resource zip name conflict: ' + subdirname)
     build_utils.ExtractAll(z, path=subdir)
     dep_subdirs.append(subdir)
   return dep_subdirs
 
 
 class _ResourceBuildContext(object):
-  """A temporary directory for packaging and compiling Android resources."""
-  def __init__(self):
+  """A temporary directory for packaging and compiling Android resources.
+
+  Args:
+    temp_dir: Optional root build directory path. If None, a temporary
+      directory will be created, and removed in Close().
+  """
+  def __init__(self, temp_dir=None):
     """Initialized the context."""
     # The top-level temporary directory.
-    self.temp_dir = tempfile.mkdtemp()
+    if temp_dir:
+      self.temp_dir = temp_dir
+      self.remove_on_exit = False
+    else:
+      self.temp_dir = tempfile.mkdtemp()
+      self.remove_on_exit = True
+
     # A location to store resources extracted form dependency zip files.
     self.deps_dir = os.path.join(self.temp_dir, 'deps')
     os.mkdir(self.deps_dir)
@@ -422,14 +571,15 @@ class _ResourceBuildContext(object):
 
   def Close(self):
     """Close the context and destroy all temporary files."""
-    shutil.rmtree(self.temp_dir)
+    if self.remove_on_exit:
+      shutil.rmtree(self.temp_dir)
 
 
 @contextlib.contextmanager
-def BuildContext():
+def BuildContext(temp_dir=None):
   """Generator for a _ResourceBuildContext instance."""
   try:
-    context = _ResourceBuildContext()
+    context = _ResourceBuildContext(temp_dir)
     yield context
   finally:
     context.Close()
@@ -449,8 +599,9 @@ def ResourceArgsParser():
 
   build_utils.AddDepfileOption(output_opts)
 
-  input_opts.add_argument('--android-sdk-jars', required=True,
-                        help='Path to the android.jar file.')
+  input_opts.add_argument('--include-resources', required=True, action="append",
+                        help='Paths to arsc resource files used to link '
+                             'against. Can be specified multiple times.')
 
   input_opts.add_argument('--aapt-path', required=True,
                          help='Path to the Android aapt tool')
@@ -479,6 +630,15 @@ def ResourceArgsParser():
            'list of resources to be included in the R.java file in the format '
            'generated by aapt.')
 
+  input_opts.add_argument(
+      '--package-name',
+      help='Package name that will be used to determine package ID.')
+
+  input_opts.add_argument(
+      '--package-name-to-id-mapping',
+      help='List containing mapping from package name to package IDs that will '
+      'be assigned.')
+
   return (parser, input_opts, output_opts)
 
 
@@ -489,7 +649,11 @@ def HandleCommonOptions(options):
     options: the result of parse_args() on the parser returned by
         ResourceArgsParser(). This function updates a few common fields.
   """
-  options.android_sdk_jars = build_utils.ParseGnList(options.android_sdk_jars)
+  options.include_resources = [build_utils.ParseGnList(r) for r in
+                               options.include_resources]
+  # Flatten list of include resources list to make it easier to use.
+  options.include_resources = [r for resources in options.include_resources
+                               for r in resources]
 
   options.dependencies_res_zips = (
       build_utils.ParseGnList(options.dependencies_res_zips))
@@ -509,3 +673,112 @@ def HandleCommonOptions(options):
 
   if not options.aapt2_path:
     options.aapt2_path = options.aapt_path + '2'
+
+
+def ParseAndroidResourceStringsFromXml(xml_data):
+  """Parse and Android xml resource file and extract strings from it.
+
+  Args:
+    xml_data: XML file data.
+  Returns:
+    A (dict, namespaces) tuple, where |dict| maps string names to their UTF-8
+    encoded value, and |namespaces| is a dictionary mapping prefixes to URLs
+    corresponding to namespaces declared in the <resources> element.
+  """
+  # NOTE: This uses regular expression matching because parsing with something
+  # like ElementTree makes it tedious to properly parse some of the structured
+  # text found in string resources, e.g.:
+  #      <string msgid="3300176832234831527" \
+  #         name="abc_shareactionprovider_share_with_application">\
+  #             "Condividi tramite <ns1:g id="APPLICATION_NAME">%s</ns1:g>"\
+  #      </string>
+  result = {}
+
+  # Find <resources> start tag and extract namespaces from it.
+  m = re.search('<resources([^>]*)>', xml_data, re.MULTILINE)
+  if not m:
+    raise Exception('<resources> start tag expected: ' + xml_data)
+  input_data = xml_data[m.end():]
+  resource_attrs = m.group(1)
+  re_namespace = re.compile('\s*(xmlns:(\w+)="([^"]+)")')
+  namespaces = {}
+  while resource_attrs:
+    m = re_namespace.match(resource_attrs)
+    if not m:
+      break
+    namespaces[m.group(2)] = m.group(3)
+    resource_attrs = resource_attrs[m.end(1):]
+
+  # Find each string element now.
+  re_string_element_start = re.compile('<string ([^>]* )?name="([^">]+)"[^>]*>')
+  re_string_element_end = re.compile('</string>')
+  while input_data:
+    m = re_string_element_start.search(input_data)
+    if not m:
+      break
+    name = m.group(2)
+    input_data = input_data[m.end():]
+    m2 = re_string_element_end.search(input_data)
+    if not m2:
+      raise Exception('Expected closing string tag: ' + input_data)
+    text = input_data[:m2.start()]
+    input_data = input_data[m2.end():]
+    if len(text) and text[0] == '"' and text[-1] == '"':
+      text = text[1:-1]
+    result[name] = text
+
+  return result, namespaces
+
+
+def GenerateAndroidResourceStringsXml(names_to_utf8_text, namespaces=None):
+  """Generate an XML text corresponding to an Android resource strings map.
+
+  Args:
+    names_to_text: A dictionary mapping resource names to localized
+      text (encoded as UTF-8).
+    namespaces: A map of namespace prefix to URL.
+  Returns:
+    New non-Unicode string containing an XML data structure describing the
+    input as an Android resource .xml file.
+  """
+  result = '<?xml version="1.0" encoding="utf-8"?>\n'
+  result += '<resources'
+  if namespaces:
+    for prefix, url in sorted(namespaces.iteritems()):
+      result += ' xmlns:%s="%s"' % (prefix, url)
+  result += '>\n'
+  if not names_to_utf8_text:
+    result += '<!-- this file intentionally empty -->\n'
+  else:
+    for name, utf8_text in sorted(names_to_utf8_text.iteritems()):
+      result += '<string name="%s">"%s"</string>\n' % (name, utf8_text)
+  result += '</resources>\n'
+  return result
+
+
+def FilterAndroidResourceStringsXml(xml_file_path, string_predicate):
+  """Remove unwanted localized strings from an Android resource .xml file.
+
+  This function takes a |string_predicate| callable object that will
+  receive a resource string name, and should return True iff the
+  corresponding <string> element should be kept in the file.
+
+  Args:
+    xml_file_path: Android resource strings xml file path.
+    string_predicate: A predicate function which will receive the string name
+      and shal
+  """
+  with open(xml_file_path) as f:
+    xml_data = f.read()
+  strings_map, namespaces = ParseAndroidResourceStringsFromXml(xml_data)
+
+  string_deletion = False
+  for name in strings_map.keys():
+    if not string_predicate(name):
+      del strings_map[name]
+      string_deletion = True
+
+  if string_deletion:
+    new_xml_data = GenerateAndroidResourceStringsXml(strings_map, namespaces)
+    with open(xml_file_path, 'wb') as f:
+      f.write(new_xml_data)
